Extract MetadataMergeStage from AstTransformsStage
The metadata merge (project + directory + document + runtime → single
flattened config) was inside AstTransformsStage::run. This is
architecturally wrong — metadata merge is a configuration resolution
step, not an AST transform. It also blocks future work: any pipeline
stage that needs merged metadata (e.g., CSS compilation) had to run
after AstTransformsStage.

This extracts the merge into its own MetadataMergeStage that runs
immediately before AstTransformsStage in all pipeline builders:

  ParseDocument → EngineExecution → MetadataMerge → AstTransforms → ...

Pure refactor — no behavior change. The same merge logic runs at the
same point in the pipeline, just in a dedicated stage.

Files changed:
- New: stages/metadata_merge.rs (stage + 12 moved tests)
- stages/ast_transforms.rs: removed merge logic and merge tests
- stages/mod.rs, stage/mod.rs: re-export MetadataMergeStage
- pipeline.rs: insert MetadataMergeStage in all 4 pipeline builders,
  update stage count assertions and doc comments
