+++ crates/quarto-core/src/stage/stages/mod.rs
@@ -12,6 +12,7 @@
 //!
 //! - [`ParseDocumentStage`] - Parse QMD content to Pandoc AST
 //! - [`EngineExecutionStage`] - Execute code cells via knitr/jupyter/markdown
+//! - [`MetadataMergeStage`] - Merge project/directory/document/runtime metadata
 //! - [`AstTransformsStage`] - Apply Quarto-specific AST transforms
 //! - [`RenderHtmlBodyStage`] - Render AST to HTML body
 //! - [`ApplyTemplateStage`] - Apply HTML template to rendered body
@@ -19,11 +20,13 @@
 mod apply_template;
 mod ast_transforms;
 mod engine_execution;
+mod metadata_merge;
 mod parse_document;
 mod render_html;
 
 pub use apply_template::{ApplyTemplateConfig, ApplyTemplateStage};
 pub use ast_transforms::AstTransformsStage;
 pub use engine_execution::EngineExecutionStage;
+pub use metadata_merge::MetadataMergeStage;
 pub use parse_document::ParseDocumentStage;
 pub use render_html::RenderHtmlBodyStage;
